# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: XAUUSD Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trades for current user
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Log a new trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Trade created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
  /trades/{id}:
    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a trade (e.g. close price, status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeUpdate"
      responses:
        "200":
          description: Trade updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── DASHBOARD ────────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    Trade:
      type: object
      required: [id, userId, tradeDate, direction, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        tradeDate:
          type: string
        balance:
          type: ["number", "null"]
        riskPct:
          type: ["number", "null"]
        entryPrice:
          type: ["number", "null"]
        slPrice:
          type: ["number", "null"]
        tpPrice:
          type: ["number", "null"]
        lotSize:
          type: ["number", "null"]
        direction:
          type: string
          enum: [Long, Short]
        status:
          type: string
          enum: [Pending, Running, TP Hit, SL Hit]
        closePrice:
          type: ["number", "null"]
        pips:
          type: ["number", "null"]
        pnl:
          type: ["number", "null"]
        tags:
          type: ["string", "null"]
          description: Comma-separated trade tags such as Scalp, Swing, or News
        session:
          type: ["string", "null"]
          enum: [Asia, London, New York]
        notes:
          type: ["string", "null"]
        lossReason:
          type: ["string", "null"]
          description: Mistake Journal — why did this losing trade fail? (FOMO, Revenge Trade, News, Wrong SL, etc.)
        createdAt:
          type: string
    TradeInput:
      type: object
      required: [tradeDate, direction]
      properties:
        tradeDate:
          type: string
        balance:
          type: ["number", "null"]
        riskPct:
          type: ["number", "null"]
        entryPrice:
          type: ["number", "null"]
        slPrice:
          type: ["number", "null"]
        tpPrice:
          type: ["number", "null"]
        lotSize:
          type: ["number", "null"]
        direction:
          type: string
          enum: [Long, Short]
        status:
          type: string
          enum: [Pending, Running, TP Hit, SL Hit]
        closePrice:
          type: ["number", "null"]
        pips:
          type: ["number", "null"]
        pnl:
          type: ["number", "null"]
        tags:
          type: ["string", "null"]
        session:
          type: ["string", "null"]
          enum: [Asia, London, New York]
        notes:
          type: ["string", "null"]
        lossReason:
          type: ["string", "null"]
          description: Mistake Journal — reason for loss
    TradeUpdate:
      type: object
      properties:
        balance:
          type: ["number", "null"]
        riskPct:
          type: ["number", "null"]
        entryPrice:
          type: ["number", "null"]
        slPrice:
          type: ["number", "null"]
        tpPrice:
          type: ["number", "null"]
        lotSize:
          type: ["number", "null"]
        direction:
          type: string
          enum: [Long, Short]
        status:
          type: string
          enum: [Pending, Running, TP Hit, SL Hit]
        closePrice:
          type: ["number", "null"]
        pips:
          type: ["number", "null"]
        pnl:
          type: ["number", "null"]
        tags:
          type: ["string", "null"]
        session:
          type: ["string", "null"]
          enum: [Asia, London, New York]
        notes:
          type: ["string", "null"]
        lossReason:
          type: ["string", "null"]
          description: Mistake Journal — reason for loss

    # DASHBOARD
...